Extract gpio_handler into separate module.
