+++ src/gpio.c
@@ -25,20 +25,6 @@
 #include "debug.h"
 #include "system.h"
 #include "peripherals.h"
-#include "stm32h7xx_hal.h"
-
-/**************************************************************************************
- * TYPEDEF
- **************************************************************************************/
-
-struct GPIO_numbers {
-  GPIO_TypeDef * port;
-  uint16_t pin;
-};
-
-struct IRQ_numbers {
-  uint16_t pin;
-};
 
 /**************************************************************************************
  * GLOBAL VARIABLES
@@ -91,14 +77,6 @@ struct IRQ_numbers IRQ_pinmap[16];
 static volatile uint16_t int_event_flags = 0;
 
 /**************************************************************************************
- * INTERNAL FUNCTION DECLARATION
- **************************************************************************************/
-
-static void gpio_disable_irq(uint8_t pin);
-static void gpio_enable_irq(uint8_t pin);
-static void gpio_set_handler(uint8_t pin);
-
-/**************************************************************************************
  * FUNCTION DEFINITION
  **************************************************************************************/
 
@@ -113,9 +91,6 @@ static void MX_GPIO_Init(void) {
   __HAL_RCC_GPIOF_CLK_ENABLE();
   __HAL_RCC_GPIOG_CLK_ENABLE();
 }
-
-#define GPIO_MODE_IN_RE         0x01   /*!< Input interrupt rising edge */
-#define GPIO_MODE_IN_FE         0x02   /*!< Input interrupt falling edge */
 
 static void handle_irq() {
   uint32_t pr = EXTI->PR1;
@@ -139,7 +114,7 @@ static void handle_irq() {
   }
 }
 
-static void gpio_disable_irq(uint8_t pin) {
+void gpio_disable_irq(uint8_t pin) {
   dbg_printf("gpio_disable_irq: pin = %x\n", pin);
   if (pin == GPIO_PIN_0) {
     HAL_NVIC_DisableIRQ(EXTI0_IRQn);
@@ -161,7 +136,7 @@ static void gpio_disable_irq(uint8_t pin) {
   }
 }
 
-static void gpio_enable_irq(uint8_t pin) {
+void gpio_enable_irq(uint8_t pin) {
   dbg_printf("gpio_enable_irq: pin = %x\n", pin);
   if (pin == GPIO_PIN_0) {
     HAL_NVIC_EnableIRQ(EXTI0_IRQn);
@@ -178,7 +153,7 @@ static void gpio_enable_irq(uint8_t pin) {
   }
 }
 
-static void gpio_set_handler(uint8_t pin)
+void gpio_set_handler(uint8_t pin)
 {
   if (pin == GPIO_PIN_0) {
     NVIC_SetVector(EXTI0_IRQn, (uint32_t)&handle_irq);
@@ -195,7 +170,8 @@ static void gpio_set_handler(uint8_t pin)
   }
 }
 
-static uint8_t GPIO_PIN_to_index(uint32_t pin) {
+uint8_t GPIO_PIN_to_index(uint32_t pin)
+{
   uint8_t index = 0;
   while (pin >>= 1) {
     index++;
@@ -203,77 +179,9 @@ static uint8_t GPIO_PIN_to_index(uint32_t pin) {
   return index;
 }
 
-int gpio_handler(uint8_t opcode, uint8_t *pdata, uint16_t size) {
-  uint16_t data = *((uint16_t*)pdata);
-  enum Opcodes_GPIO action = opcode;
-
-  uint8_t value = (data & 0xFF00) >> 8;
-  uint8_t index = data & 0xFF;
-
-  GPIO_InitTypeDef GPIO_InitStruct = {0};
-  uint8_t response[2];
-
-  switch (action) {
-    case CONFIGURE:
-      GPIO_InitStruct.Pin = GPIO_pinmap[index].pin;
-      GPIO_InitStruct.Mode = value;
-      GPIO_InitStruct.Pull = GPIO_PULLUP;
-      GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-      HAL_GPIO_Init(GPIO_pinmap[index].port, &GPIO_InitStruct);
-      dbg_printf("GPIO%d: CONFIGURE %d\n", index, value);
-      break;
-    case IRQ_TYPE:
-      GPIO_InitStruct.Pin = GPIO_pinmap[index].pin;
-
-      if      (value == GPIO_MODE_IN_RE) GPIO_InitStruct.Mode = GPIO_MODE_IT_RISING;
-      else if (value == GPIO_MODE_IN_FE) GPIO_InitStruct.Mode = GPIO_MODE_IT_FALLING;
-      else                               return 0;
-
-      GPIO_InitStruct.Pull = GPIO_PULLUP;
-      GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-      HAL_GPIO_Init(GPIO_pinmap[index].port, &GPIO_InitStruct);
-      IRQ_pinmap[GPIO_PIN_to_index(GPIO_InitStruct.Pin)].pin = index;
-      dbg_printf("GPIO%d: IRQ_TYPE %d\n", index, value);
-      break;
-    case IRQ_ENABLE:
-      dbg_printf("GPIO%d: IRQ_ENABLE %d\n", index, value);
-      if (value == 1) {
-        gpio_set_handler(GPIO_pinmap[index].pin);
-        gpio_enable_irq(GPIO_pinmap[index].pin);
-      } else {
-        gpio_disable_irq(GPIO_pinmap[index].pin);
-      }
-      break;
-    case WRITE:
-      HAL_GPIO_WritePin(GPIO_pinmap[index].port, GPIO_pinmap[index].pin, value);
-      dbg_printf("GPIO%d: WRITE %d\n", index, value);
-      break;
-    case READ:
-      response[0] = index;
-      response[1] = HAL_GPIO_ReadPin(GPIO_pinmap[index].port, GPIO_pinmap[index].pin);
-      dbg_printf("GPIO%d: READ %d\n", index, response[1]);
-      return enqueue_packet(PERIPH_GPIO, opcode, sizeof(response), &response);
-      break;
-    case IRQ_SIGNAL:
-      // do nothing;
-      dbg_printf("GPIO%d: IRQ_SIGNAL %d\n", index, value);
-      break;
-    case IRQ_ACK:
-      dbg_printf("GPIO%d: IRQ_ACK %d\n", index, value);
-      /* Re-enable the interrupt that was disabled within
-       * handle_irq to prevent firing of another interrupt
-       * until this one has been signaled to the application.
-       */
-      gpio_enable_irq(GPIO_pinmap[index].pin);
-      break;
-  }
-  return 0;
-}
-
-void gpio_init() {
+void gpio_init()
+{
   MX_GPIO_Init();
-
-  register_peripheral_callback(PERIPH_GPIO, &gpio_handler);
 }
 
 void gpio_set_initial_config() {
